fix(agent-chat): expose chat types through contracts
