+++ apps/desktop/src/renderer/src/agent-chat/messages/system-message.tsx
@@ -1,4 +1,4 @@
-import type { Message } from '@main/agent/storage/types'
+import type { Message } from '@memry/contracts/ipc-agent'
 
 import { formatUnknown } from './format'
